[Object Counters] Introduce Feature Gates For The Various Object Counters [1.7.x] (#6981)
https://iguazio.atlassian.net/browse/ML-8556

Our current counter logic is naive and can cause unwanted database locks
which prevent backups from completing. We are working on a more robust
counter system, though for now we are introducing feature gates to
disable the problematic counter queries on very heavy systems.

Remove `_calculate_functions_counters` as it is not used anywhere and is
private and internal.