[Model Monitoring] Create the schemas for all V3IO tables during deployment (#7420)
Currently, we create only 2 out of the 5 V3IO tables during model
monitoring infra deployment (`metrics` and `app-results`). The other 3
tables are generated only when the first relevant event is pushed to
them. For example, the `errors` table is created only when the first
error event occurs. As a result, when `framesd` tries to query these
tables (e.g., during `list_model_endpoints` in each controller process),
it gets an error about a missing TSDB schema file.

In this PR, we address this by generating all of the schemas during
deployment. This ensures the schemas are available in advance and
prevents the repeating `No TSDB schema file found` error in `framesd`
logs."

https://iguazio.atlassian.net/browse/ML-9542