[Nuclio] Update client-side with server-side spec changes after deployment (#7631)
This PR fixes an issue where replica values (`min_replicas` and
`max_replicas`) in the function spec were incorrect after deployment.

The cause of this is that after deploying a function, we call
`_enrich_command_from_status`, which saved the outdated client-side
function object. This overwrote server-side updates, causing
inconsistencies in the function's spec.

So we now update the function’s spec from the server-side deployment
response (similar to how we already update the status). This ensures
that all server-side changes, including replica settings, are correctly
reflected on the client side.

Jira:
https://iguazio.atlassian.net/browse/ML-9725