[Nuclio] Fix spec enrichment if base_spec is not empty (#8081)
Jira - https://iguazio.atlassian.net/browse/ML-10206

Test doesn't work without fix. In general, it is better to use
`set_function` in tests than `code_to_function` because their behaviour
is different.

The root cause is that if base_spec is not empty, then we do not get a
basic config -
https://github.com/nuclio/nuclio-jupyter/blob/21b122bf97b86e1926fd0a74b951f377da42f8bb/nuclio/config.py#L37-L54
and as a result, we fail on not having labels key here
https://github.com/nuclio/nuclio-jupyter/blob/21b122bf97b86e1926fd0a74b951f377da42f8bb/nuclio/config.py#L391.

tested on a system:

![image](https://github.com/user-attachments/assets/03e404f5-4420-4615-a32b-12a55429a19e)

![image](https://github.com/user-attachments/assets/6f205d06-3816-4a5c-b162-327889a29da1)