[Deprecations] Change functionality of `allow_cross_project` in `mlrun.load_project` (#8085)
Project name differs from the name specified in the context's project
YAML.
This functionality is no longer supported. If you want to enable this
behavior, you can take one of the following actions:

1. Set `allow_cross_project=True` when loading the project.
2. Delete the existing project YAML, or ensure its name field matches
the actual project name.
3. Use a different project context directory.

Previously, when `allow_cross_project` was not set (None), it implicitly
behaved as if it were True, with a warning. Now, if you want this
functionality, you must explicitly set `allow_cross_project=True`.

Also, added support for `allow_cross_project` in the CLI with default as
true.

https://iguazio.atlassian.net/browse/ML-10062