[Workflows] Add allow cross project flag to workflow runner  (#8143)
This PR sets allow_cross_project=True when calling `load_project` and
`get_or_create_project` in the workflow-runner logic. This change
prevents failures when the project name in the project.yaml differs from
the one passed to the function, which is a common scenario in remote
workflows.

Without this flag, the remote runner fail with a project name mismatch
error, as shown in the attached example. We already rely on this
behavior in our system tests for remote workflows, where the project
name is intentionally different.


![image](https://github.com/user-attachments/assets/3229ca52-b38d-4164-b2d2-0007cec5ae4f)